# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 校园寻物 Campus Finder API
  version: 0.1.0
  description: |
    校园失物招领系统前后端接口契约初版。

    约定：
    - 成功响应统一使用 code=0，失败响应使用非 0 业务错误码。
    - 一个业务错误码只对应一个固定 msg；HTTP 状态码仍用于表达协议层结果。
    - code=10000 表示服务器内部错误（未预期异常的兜底），不属于业务错误码序列。
    - code=10004：管理员账号不允许自我注销。
    - 登录接口返回 JWT，后续请求使用 Authorization: Bearer <token>。
    - 普通用户只能操作自己的发布和申请；角色权限必须由后端校验。
    - 非管理员访问管理员接口返回 10010；lost_admin 访问仅 system_admin 可用的接口返回 10011。
    - 普通列表只展示 reviewStatus=approved 的物品。
  contact:
    name: Campus Finder Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 线上环境占位地址，部署后替换
...
paths:
...
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前登录用户
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 当前用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/profile:
    patch:
      tags: [Auth]
      summary: 修改个人信息
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/InvalidParameters'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/account:
    delete:
      tags: [Auth]
      summary: 注销账号
      description: 软删除当前账号并下架名下全部物品，待审核的认领申请自动取消，学号释放后可重新注册。管理员账号不允许注销。注销前需验证当前密码。
      operationId: deleteAccount
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '200':
          description: 注销成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/InvalidParameters'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/AdminAccountUndeletable'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    InvalidParameters:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 10001
            msg: 参数错误
            data: null
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 10002
            msg: 未登录或登录已过期
            data: null
...
    AdminAccountUndeletable:
      description: 管理员账号不允许注销
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 10004
            msg: 管理员账号不允许注销
            data: null
...
  schemas:
    UserRole:
      type: string
      enum: [student, lost_admin, system_admin]
      description: 用户角色
    UserStatus:
      type: string
      enum: [active, disabled]
...
    User:
      type: object
      required: [id, username, name, role, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          description: 学号，12 位纯数字
          example: '302026315155'
        name:
          type: string
          example: 张三
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
          example: '2026-09-14T08:00:00+08:00'
...
    UpdateProfileRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: 李四
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          format: password
          example: password123
...
    ErrorResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          description: 业务错误码，0 表示成功；10000 为服务器内部错误兜底码，其余按模块分段（100xx 通用、120xx 物品、130xx 认领、140xx 用户、150xx 公告、160xx 图片）
          example: 10001
        msg:
          type: string
          description: 与 code 一一对应的固定错误消息
          example: 参数错误
        data:
          type: object
          nullable: true
          description: 错误时固定为 null
          example: null
    EmptyDataResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          example: 0
        msg:
          type: string
          example: success
        data:
          type: object
          nullable: true
          example: null
...
    UserResponse:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          example: 0
        msg:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/User'
...